Avoid last task override
Thanks for the example, it has been very useful for me to understand how Truffle works.

I noticed that in NoteContract.sol, the statement "noteCount--" when removing a task from listTile different from the last task, after doing fetchNote(), the last task is avoided because its id was not taken into account. Which causes it to no longer appear in the UI and later been replaced by another task when creating a new one.

To avoid this, it will be necessary to remove "noteCount--", in addition to the fact that there will no longer be id repetition.